# local-shared-data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnyReason API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/episodes/{episode_id}/asset-bindings": {
      "get": {
        "tags": [
          "asset_bindings"
        ],
        "summary": "List Episode Asset Bindings",
        "operationId": "list_episode_asset_bindings",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBase_list_AssetBindingBrief__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storyboards/{storyboard_id}/asset-bindings": {
      "get": {
        "tags": [
          "asset_bindings"
        ],
        "summary": "List Storyboard Asset Bindings",
        "operationId": "list_storyboard_asset_bindings",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "storyboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Storyboard Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBase_StoryboardAssetBindingsResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "asset_bindings"
        ],
        "summary": "Upsert Storyboard Asset Binding",
        "operationId": "upsert_storyboard_asset_binding",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "storyboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Storyboard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetBindingCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBase_AssetBindingBrief_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asset-bindings/{binding_id}": {
      "delete": {
        "tags": [
          "asset_bindings"
        ],
        "summary": "Delete Asset Binding",
        "operationId": "delete_asset_binding",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "binding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Binding Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBase_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/storyboards/{storyboard_id}/ai/storyboard/prompt-preview": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Preview Storyboard Prompt",
        "operationId": "preview_storyboard_prompt",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "storyboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Storyboard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AISceneStoryboardPromptPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBase_AISceneStoryboardPromptPreviewResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storyboards/{storyboard_id}/ai/storyboard/preview": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Preview Storyboard",
        "operationId": "preview_storyboard",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "storyboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Storyboard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AISceneStoryboardPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBase_AISceneStoryboardPreviewResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storyboards/{storyboard_id}/ai/storyboard/apply": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Apply Storyboard",
        "operationId": "apply_storyboard",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "storyboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Storyboard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AISceneStoryboardApplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBase_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ai/video/generate": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Ai Generate Video",
        "operationId": "ai_generate_video",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIVideoGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBase_AIVideoGenerateResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AISceneStoryboardApplyRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "replace",
              "append"
            ],
            "title": "Mode",
            "default": "replace"
          },
          "shots": {
            "items": {
              "$ref": "#/components/schemas/AIShotDraft"
            },
            "type": "array",
            "title": "Shots"
          }
        },
        "type": "object",
        "title": "AISceneStoryboardApplyRequest"
      },
      "AISceneStoryboardPreviewRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "prompt_template": {
            "type": "string",
            "title": "Prompt Template"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          }
        },
        "type": "object",
        "required": [
          "model",
          "prompt_template"
        ],
        "title": "AISceneStoryboardPreviewRequest"
      },
      "AISceneStoryboardPreviewResponse": {
        "properties": {
          "final_prompt": {
            "type": "string",
            "title": "Final Prompt"
          },
          "raw_text": {
            "type": "string",
            "title": "Raw Text"
          },
          "shots": {
            "items": {
              "$ref": "#/components/schemas/AIShotDraft"
            },
            "type": "array",
            "title": "Shots"
          }
        },
        "type": "object",
        "required": [
          "final_prompt",
          "raw_text"
        ],
        "title": "AISceneStoryboardPreviewResponse"
      },
      "AISceneStoryboardPromptPreviewRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "prompt_template": {
            "type": "string",
            "title": "Prompt Template"
          }
        },
        "type": "object",
        "required": [
          "model",
          "prompt_template"
        ],
        "title": "AISceneStoryboardPromptPreviewRequest"
      },
      "AISceneStoryboardPromptPreviewResponse": {
        "properties": {
          "final_prompt": {
            "type": "string",
            "title": "Final Prompt"
          }
        },
        "type": "object",
        "required": [
          "final_prompt"
        ],
        "title": "AISceneStoryboardPromptPreviewResponse"
      },
...
      "AIVideoGenerateRequest": {
        "properties": {
          "binding_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Binding Key",
            "default": "video"
          },
          "model_config_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Config Id"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          },
          "duration": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Duration",
            "default": 5
          },
          "aspect_ratio": {
            "type": "string",
            "title": "Aspect Ratio",
            "default": "16:9"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "AIVideoGenerateRequest"
      },
      "AIVideoGenerateResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "raw": {
            "type": "object",
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "AIVideoGenerateResponse"
      },
...
      "AssetBindingBrief": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "asset_entity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asset Entity Id"
          },
          "asset_variant_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Variant Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "variant_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Code"
          },
          "stage_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Tag"
          },
          "age_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Range"
          }
        },
        "type": "object",
        "required": [
          "id",
          "asset_entity_id",
          "name",
          "type"
        ],
        "title": "AssetBindingBrief"
      },
      "AssetBindingCreateRequest": {
        "properties": {
          "asset_entity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asset Entity Id"
          },
          "asset_variant_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Variant Id"
          }
        },
        "type": "object",
        "required": [
          "asset_entity_id"
        ],
        "title": "AssetBindingCreateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponseBase_AISceneStoryboardPreviewResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "OK"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AISceneStoryboardPreviewResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ResponseBase[AISceneStoryboardPreviewResponse]"
      },
      "ResponseBase_AISceneStoryboardPromptPreviewResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "OK"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AISceneStoryboardPromptPreviewResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ResponseBase[AISceneStoryboardPromptPreviewResponse]"
      },
...
      "ResponseBase_AIVideoGenerateResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "OK"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIVideoGenerateResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ResponseBase[AIVideoGenerateResponse]"
      },
...
      "ResponseBase_AssetBindingBrief_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "OK"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetBindingBrief"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ResponseBase[AssetBindingBrief]"
      },
...
      "ResponseBase_StoryboardAssetBindingsResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "OK"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoryboardAssetBindingsResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ResponseBase[StoryboardAssetBindingsResponse]"
      },
...
      "ResponseBase_dict_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "OK"
          },
          "data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ResponseBase[dict]"
      },
...
      "ResponseBase_list_AssetBindingBrief__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "OK"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AssetBindingBrief"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ResponseBase[list[AssetBindingBrief]]"
      },
...
      "StoryboardAssetBindingsResponse": {
        "properties": {
          "storyboard_id": {
            "type": "string",
            "format": "uuid",
            "title": "Storyboard Id"
          },
          "bindings": {
            "items": {
              "$ref": "#/components/schemas/AssetBindingBrief"
            },
            "type": "array",
            "title": "Bindings"
          }
        },
        "type": "object",
        "required": [
          "storyboard_id"
        ],
        "title": "StoryboardAssetBindingsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}